Former-commit-id: 2e1c7feba22f02ed13b38083bc1f35ffe7c851ec [formerly ae3139a9ca79699c699a715d74d681d756643730] [formerly 39220f5d350851fbc0eb34bc3f46d7b88cb566c8 [formerly fa29119f9d336a6d7e539a31f80640e852837656]] [formerly b7e09d82e4834691c2b6d341e293ebb6b194a774 [formerly e5f23ee921a8922943772c7e3dc59c44fdec28a1]]
Former-commit-id: b7e09d82e4834691c2b6d341e293ebb6b194a774
Former-commit-id: ca39a8e98f733d58c84f06b3a88ac558cfccbfd3 [formerly 97fe8f8a7b2962e98f50432303acaecca95b4a66]
Former-commit-id: 7da4c798a9e9366c607f77c830b0dc1fb3b1a2d1

+++ templating/templates/utilityFunctions.h
@@ -1,5 +1,6 @@
 #include <stdint.h>
 #include <signal.h>
+#include <stdlib.h>
 
 #define __GNU_SOURCE
 
@@ -26,7 +27,7 @@ void die(char *errorStr);
  * create a signal handler that handles signal signum and runs the function *psh
  * when signum is raised
  */
-void open_shared_mem(uint8_t **ppmem, const char *pName, int numBytes, int shm_flags, int mmap_flags);
+void open_shared_mem(uint8_t **ppmem, const char *pName, size_t numBytes, int shm_flags, int mmap_flags);
 
 /*
  * open a shared memory block with:
